client/requirements.txt
grpcio
grpcio-tools
pytermgui
